Completed most of migration to FastAPI
