+++ FSBS/API/endpoints/auth_endpoints.py
@@ -2,7 +2,7 @@ import json
 from flask import request, Response
 
 from FSBS.API.extensions import db, bcrypt
-from FSBS.API.models import User, Transaction
+from FSBS.API.database.models import User, Transaction
 from FSBS.API.responses import errors
 from FSBS.API.database import validation
