+++ FSBS/API/endpoints/transaction_endpoints.py
@@ -4,7 +4,7 @@ from FSBS.API.decorators import user_endpoint
 from FSBS.API.extensions import db
 
 from flask import request, Response
-from FSBS.API.models import User, Transaction
+from FSBS.API.database.models import User, Transaction
 from FSBS.API.responses import errors
