+++ requirements.txt
@@ -24,3 +24,5 @@ six==1.16.0
 SQLAlchemy==1.4.42
 urllib3==1.26.12
 Werkzeug==2.2.2
+
+fastapi~=0.88.0
